libs/core/pyproject.toml
[project]
name = "kamas-core"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "attrs>=25.4.0",
    "metatrader5>=5.0.5430",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
]
